Stuff generates, yay!
Might not be 100% correct, but I am seeing generation. Think that Diorite and Granite need to be added to the config.

Forced config save command needs implemented. Might be good to add a generic for doing it anyway. Issue with that is
that I don't believe we actually keep enough data around to reconstitute the original files in their original format.

Will discuss with @jriwanek at a later point.