Refactor to split out to individual spawn generator calls.
Preparation for the 1.16 port - this splits things so there is no longer a single "mega generator",
instead we have a separate generator per spawn entry.

Next step will be fixing some bits of the generators to get them hitting the chunk boundaries, as
the last changes to halt cascading chunk gen caused a "gridding" of generation rather than a nice,
clean generation.